feat: use og api route again and remove opengraph-image
