+++ src/utils/og.ts
@@ -22,3 +22,12 @@ export const buildOgImageUrl = (
     return `${baseUrl}/api/og`;
   }
 };
+
+export const config = {
+  runtime: 'edge',
+  size: {
+    width: 1200,
+    height: 630,
+  },
+  contentType: 'image/png',
+};
